Handling-products-response

+++ app/src/main/java/com/example/dummyproducts/models/Rating.kt
@@ -1,6 +1,6 @@
 package com.example.dummyproducts.models
 
 data class Rating(
-    val count: Int,
+    val count: Double,
     val rate: Double
 )
